shift around to remove ciruclar dependencies
