+++ extension/src/externalresources/BaseAppTranslationFiles.ts
@@ -1,7 +1,5 @@
-import { readdirSync, readFileSync, unlink } from "fs";
+import * as fs from "fs";
 import { resolve, basename } from "path";
-import * as LanguageFunctions from "../LanguageFunctions";
-import * as SettingsLoader from "../Settings/SettingsLoader";
 
 import { BlobContainer } from "./ExternalResources";
 
@@ -48,7 +46,7 @@ baseAppTranslationFiles.addBlob("sv-se.json");
  */
 export function localBaseAppTranslationFiles(): Map<string, string> {
   const files: Map<string, string> = new Map<string, string>();
-  readdirSync(__dirname)
+  fs.readdirSync(__dirname)
     .filter((a) => a.endsWith(".json"))
     .forEach((file) => {
       if (file.match(languageCodeJsonRE)) {
@@ -56,41 +54,4 @@ export function localBaseAppTranslationFiles(): Map<string, string> {
       }
     });
   return files;
-}
-
-export async function validateLocalBaseAppTranslationFiles(
-  printToConsole = false
-): Promise<number> {
-  const targetLanguageCodes = LanguageFunctions.existingTargetLanguageCodes(
-    SettingsLoader.getSettings(),
-    SettingsLoader.getAppManifest()
-  );
-  const invalidFiles = [];
-  // For optimisation we only check files if there is a target xlf with matching language Code
-  const localFiles = localBaseAppTranslationFiles();
-  for (const k of localFiles.keys()) {
-    if (!targetLanguageCodes?.includes(k.replace(".json", ""))) {
-      localFiles.delete(k);
-    }
-  }
-  if (localFiles.size === 0) {
-    return localFiles.size;
-  }
-
-  for (const file of localFiles.entries()) {
-    try {
-      JSON.parse(readFileSync(file[1], "utf8"));
-    } catch (error) {
-      invalidFiles.push(file[1]);
-    }
-  }
-  invalidFiles.forEach((f) => {
-    unlink(f, () => {
-      // async unlink requires callback
-    });
-    if (printToConsole) {
-      console.log(`NAB AL Tools: Removed invalid translation map at: ${f}`);
-    }
-  });
-  return invalidFiles.length;
 }
